+++ examples/modbus/position.py
@@ -5,8 +5,8 @@ from edcon.utils.logging import Logging
 # Enable loglevel info
 Logging()
 
-edrive = ComModbus('192.168.0.1')
-with MotionHandler(edrive) as mot:
+com = ComModbus('192.168.0.1')
+with MotionHandler(com) as mot:
     mot.acknowledge_faults()
     mot.enable_powerstage()
     mot.referencing_task()