+++ src/edcon/cli/tg102.py
@@ -13,9 +13,9 @@ def add_tg102_parser(subparsers):
                               help='Moment reduction to use in percent (default: %(default)s).')
 
 
-def tg102_func(edrive, args):
+def tg102_func(com, args):
     """Executes subcommand based on provided arguments"""
-    with Telegram102Handler(edrive) as tg102:
+    with Telegram102Handler(com) as tg102:
         tg102.telegram.momred.value = round(
             16384.0 * float(args.moment_reduction) / 100.0)